+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # master
 
+* Fix Client#room not being reset when leaving a room
 * Fix roomid / section not being specify consistently in room APIs
 
 ## 0.1.3 / 2012-12-25